==> src/2_remove_not_swapped_sentence_from_the_text.ipynb <==
import os
import pandas as pd
import shutil
import re

## T# Nornalization

def renumber_ann_file(file_path):
    with open(file_path, "r", encoding="utf-8") as f:
        lines = f.readlines()
    
    t_annotations = [line for line in lines if line.startswith("T")]
    others = [line for line in lines if not line.startswith("T")]
    
    new_t_annotations = []
    t_map = {}
    for i, line in enumerate(t_annotations, start=1):
        old_t_num = re.match(r"T\d+", line).group()
        new_t_num = f"T{i}"
        t_map[old_t_num] = new_t_num
        new_t_annotations.append(re.sub(r"T\d+", new_t_num, line, count=1))
    
    new_other_annotations = []
    for line in others:
        for old_t, new_t in t_map.items():
            line = line.replace(old_t, new_t)
        new_other_annotations.append(line)
    
    with open(file_path, "w", encoding="utf-8") as f:
        f.writelines(new_t_annotations + new_other_annotations)

def renumber_directory(directory):
    for filename in os.listdir(directory):
        if filename.endswith(".ann"):
            file_path = os.path.join(directory, filename)
            renumber_ann_file(file_path)
            print(f"Processed: {filename}")

## Filtered corpora to leave only sentences with JOB entities

def filter_ann_file(file_path):
    with open(file_path, "r", encoding="utf-8") as f:
        lines = f.readlines()
    
    filtered_lines = [line for line in lines if re.match(r"T\d+\t(JOB|PERS)\b", line)]
    
    with open(file_path, "w", encoding="utf-8") as f:
        f.writelines(filtered_lines)
    
    print(f"Filtered file updated: {file_path}")

def process_directory(directory):
    for filename in os.listdir(directory):
        if filename.endswith(".ann"):
            file_path = os.path.join(directory, filename)
            filter_ann_file(file_path)

# source_directory = "/Users/linndfors/study/diploma/test_change_dir/ng_changed"
source_directory = "/Users/linndfors/study/diploma/test_change_dir/bruk_changed"

process_directory(source_directory)

renumber_directory(source_directory)

renaming

def rename_files(directory):
    for filename in os.listdir(directory):
        if "_1_filtered.ann" in filename:
            new_filename = filename.replace("_1_filtered.ann", "_1.ann")
            old_path = os.path.join(directory, filename)
            new_path = os.path.join(directory, new_filename)
            os.rename(old_path, new_path)
        elif "_1_filtered.txt" in filename:
            new_filename = filename.replace("_1_filtered.txt", "_1.txt")
            old_path = os.path.join(directory, filename)
            new_path = os.path.join(directory, new_filename)
            os.rename(old_path, new_path)


directory = "/Users/linndfors/study/diploma/test_change_dir/bruk_changed"
# directory = "/Users/linndfors/study/diploma/test_change_dir/ng_changed"


rename_files(directory)
