==> src/5_creation_of_parallel_dataset_for_llm.ipynb <==
import pandas as pd
import json
import ast
import numpy as np
from sklearn.model_selection import train_test_split



## Extract parallel sentences

parallel_dataset_bruk_dataset = pd.read_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/bruk_parallel.csv")
parallel_dataset_ng_dataset = pd.read_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/ng_parallel.csv")

bruk_sentence_and_ann_df = parallel_dataset_bruk_dataset[['original_sentence', 'changed_sentence']]
ng_sentence_and_ann_df = parallel_dataset_ng_dataset[['original_sentence', 'changed_sentence']]

ng_sentence_and_ann_df.head(20)

bruk_sentence_and_ann_df[35:60].to_csv("sample_of_parallel_dataset.csv", index=0)

parallel_df = pd.concat([ng_sentence_and_ann_df, bruk_sentence_and_ann_df])
parallel_df.reset_index(drop=True, inplace=True)

parallel_df.shape

parallel_df.head()

parallel_df.to_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/parallel_df.csv", index=0)

parallel_df = parallel_df.rename(columns={'original_sentence': "inputs", 'changed_sentence': 'targets'})

parallel_df = parallel_df.sample(frac=1, random_state=42).reset_index(drop=True)

swap_mask = np.random.choice([True, False], size=len(parallel_df))

inputs_copy = parallel_df['inputs'].copy()
targets_copy = parallel_df['targets'].copy()

parallel_df['inputs'] = np.where(swap_mask, targets_copy, inputs_copy)
parallel_df['targets'] = np.where(swap_mask, inputs_copy, targets_copy)

parallel_df.head()

parallel_df['inputs'] = parallel_df.apply(lambda x: "Перефразуй це речення, змінивши гендерні сутності на протилежні (чоловічий <-> жіночий): " + x['inputs'], axis=1)

parallel_df.head()

## Add word pairs (male-female)

gender_dict_df = pd.read_csv("/Users/linndfors/study/diploma/uk-gender-word-mapper/gender_pairs_dictionary.csv")

gender_dict_df = gender_dict_df.rename(columns={'male': "inputs", 'female': 'targets'})

df = gender_dict_df.sample(frac=1, random_state=42).reset_index(drop=True)

swap_mask = np.random.choice([True, False], size=len(df))

targets_copy = df['targets'].copy()

df['inputs'] = np.where(swap_mask, targets_copy, inputs_copy)
df['targets'] = np.where(swap_mask, inputs_copy, targets_copy)

df

df['inputs'] = df.apply(lambda x: "Перефразуй це слово, змінивши його гендер на протилежний (чоловічий <-> жіночий): " + x['inputs'], axis=1)

df.head()

parallel_df_enhanced = pd.concat([parallel_df, df])

parallel_df_enhanced.head()

shuffled_df = parallel_df_enhanced.sample(frac=1, random_state=42).reset_index(drop=True)

shuffled_df.head(10)

shuffled_df.to_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/parallel_for_llm_df.csv", index=0)

### Split

train_df, test_df = train_test_split(shuffled_df, test_size=0.2, random_state=42)

print("Train Set:")
print(len(train_df))
print("\nTest Set:")
print(len(test_df))

train_df.to_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/train_parallel_for_llm_df.csv", index=0)
test_df.to_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/test_parallel_for_llm_df.csv", index=0)

train_df.head()

test_df.head()

gender_dict_df

train_df = pd.read_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/train_parallel_for_llm_df.csv")
test_df = pd.read_csv("/Users/linndfors/study/diploma/ner_for_fem/data/csv_files_with_par_sentences/test_parallel_for_llm_df.csv")

train_df['inputs'][0]

test_df

### Count distribution of gender in each "inputs" column

female, male, total_words = 0, 0, 0

for x, row in train_df.iterrows():
    if "Перефразуй це слово, змінивши його гендер на протилежний (чоловічий <-> жіночий): " in row['inputs']:
        word = row['inputs'].split("Перефразуй це слово, змінивши його гендер на протилежний (чоловічий <-> жіночий): ")[1]
        total_words += 1
        if word in list(gender_dict_df['inputs'].values):
            male += 1
        elif word in list(gender_dict_df['targets'].values):
            female += 1
        else:
            print("error:", row['inputs'])
            break

print(female, male, total_words)

46200/893

43100/893

female, male, total_words = 0, 0, 0

for x, row in test_df.iterrows():
    if "Перефразуй це слово, змінивши його гендер на протилежний (чоловічий <-> жіночий): " in row['inputs']:
        word = row['inputs'].split("Перефразуй це слово, змінивши його гендер на протилежний (чоловічий <-> жіночий): ")[1]
        total_words += 1
        if word in list(gender_dict_df['inputs'].values):
            male += 1
        elif word in list(gender_dict_df['targets'].values):
            female += 1
        else:
            print("error:", row['inputs'])
            break

print(female, male, total_words)

9400/208

11400/208
